Authorization finished

+++ routes/campgrounds.js
@@ -1,25 +1,12 @@
 const express = require('express');
 const router = express.Router();
 const catchAsync = require('../utils/catchAsync')
-const ExpressError = require('../utils/ExpressError')
-const Campground = require('../models/campground');
-const {campgroundSchema ,reviewSchema} = require('../schemas.js');
-const {isLoggedIn} = require('../middleware');
 
-const validateCampground = (req,res,next) =>{
-   
-    const {error} = campgroundSchema.validate(req.body);
-    if(error){   
-        const msg = error.details.map(el => el.message).join(',')
-        throw new ExpressError(msg,400)
-        // throw new ExpressError(result.error.details,400)
-    }
-    else{
-        
-        next();
-    }
-    
-}
+const Campground = require('../models/campground');
+
+const {isLoggedIn , isAuthor ,validateCampground} = require('../middleware');
+
+
 
 router.get('/', catchAsync(async(req,res)=>{
     const campgrounds = await Campground.find({});
@@ -42,32 +29,47 @@ router.post('/',isLoggedIn, validateCampground, catchAsync(async(req,res,next)=>
     res.redirect(`/campgrounds/${campground._id}`)
 }))
 router.get('/:id', catchAsync(async(req,res)=>{
-   const campground = await Campground.findById(req.params.id).populate('reviews').populate('author');
+   const campground = await Campground.findById(req.params.id).populate({
+       path:'reviews',
+       populate:{
+        path:'author',
+       }
+}).populate('author');
+console.log(campground);
    if(!campground){
        req.flash('error', 'Cannot find that campground')
       return res.redirect('/campgrounds')
    }
-console.log(campground);
+
     res.render('campgrounds/show', {campground })
 }));
-router.get('/:id/edit',isLoggedIn, catchAsync(async(req,res)=>{
+router.get('/:id/edit',isLoggedIn,isAuthor, catchAsync(async(req,res)=>{
    const campground = await Campground.findById(req.params.id)
-   if(!campground){
-    req.flash('error', 'Cannot find that campground')
-   return res.redirect('/campgrounds')
-}
+    if(!campground){
+        req.flash('error', 'Cannot find that campground')
+       return res.redirect('/campgrounds')
+    }
+   
+   
    res.render('campgrounds/edit', {campground});
 }))
 
 router.put('/:id', isLoggedIn,validateCampground,catchAsync(async (req,res)=>{
     const {id} = req.params;
-   const campground= await Campground.findByIdAndUpdate(id, {...req.body.campground});
-   req.flash('success', "Successfully updated Campground")
+   
+   const campground = await Campground.findByIdAndUpdate(id, {...req.body.campground})
+    req.flash('success', "Successfully updated Campground")
    res.redirect(`/campgrounds/${campground._id}`)
 }))
 
 router.delete('/:id',isLoggedIn, catchAsync(async(req,res)=>{
-     const {id} = req.params;
+    const {id} = req.params;
+    const campground = await Campground.findById(id);
+  if(!campground.author.equals(req.user._id)) {
+       req.flash('error', 'You do not have permission to do that');
+      return res.redirect(`/campgrounds/${id}`)
+   } 
+    
      await Campground.findByIdAndDelete(id);
      req.flash('success', 'Succesfully deleted Campground')
